Move normalize helper out of plexvfs
Doesn't belong to plexvfs, it's more like generic helper
